requirements.txt
pyTelegramBotAPI==4.15.4
requests==2.31.0
python-dotenv==1.0.1
flask==3.0.1
psutil==5.9.8
yt-dlp